feat(#86): 투표 생성 화면 연결 - 투표 리스트 보이는 화면에서 추가 눌렀을 때 이동하도록 구현

+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/navigation/AppNavigation.kt
@@ -67,7 +67,7 @@ object AppNavigation {
             data object Management : Studies
 
             @Serializable
-            data class MakeVote(val studyGroupId: Int) : Studies
+            data class MakeVote(val studyGroupId: Long) : Studies
 
             @Serializable
             data class Record(val groupId: Long, val memberId: Long) : Studies
